requirements.txt
pandas
mysql-connector-python
streamlit
matplotlib